docs: update changelog for v3.1.0 release
- Add comprehensive documentation and test suite
- Fix SceneMask IEnumerable implementation
- Improve InactivityTimer with UniTask
- Standardize test naming and organization

+++ Runtime/InactivityTimer.cs
@@ -16,7 +16,7 @@ namespace CCC.Runtime
 
 		[Tooltip("Time in seconds before the inactivity action is triggered")]
 		[field: SerializeField]
-		public float TimeBeforeAction { get; set; } = 60;
+		public uint TimeBeforeAction { get; set; } = 60;
 
 		#endregion
